Make round-tripping Decode.Value possible
This change introduces the Encode.value function to encode a
Decode.Value into a string. This also necessitated a type name change
from Encode.Value (which was a different type from Decode.Value) to
Encode.Encoder. This better captures the purpose of Encode.Encoder
and does not lead to confusing name clashes.

This is a major change so version is bumped to 2.0.0.

This also closes #21